refactor: extract shared fuzzy filter utilities for list components
