+++ src/tui/components/filterable-select-list.ts
@@ -8,72 +8,7 @@ import {
 } from "@mariozechner/pi-tui";
 import type { Component } from "@mariozechner/pi-tui";
 import chalk from "chalk";
-
-/**
- * Fuzzy match with pre-lowercased inputs (avoids toLowerCase on every keystroke).
- * Returns score (lower = better) or null if no match.
- */
-function fuzzyMatchLower(queryLower: string, textLower: string): number | null {
-	if (queryLower.length === 0) return 0;
-	if (queryLower.length > textLower.length) return null;
-
-	let queryIndex = 0;
-	let score = 0;
-	let lastMatchIndex = -1;
-	let consecutiveMatches = 0;
-
-	for (let i = 0; i < textLower.length && queryIndex < queryLower.length; i++) {
-		if (textLower[i] === queryLower[queryIndex]) {
-			const isWordBoundary = i === 0 || /[\s\-_./:]/.test(textLower[i - 1]);
-			if (lastMatchIndex === i - 1) {
-				consecutiveMatches++;
-				score -= consecutiveMatches * 5;
-			} else {
-				consecutiveMatches = 0;
-				if (lastMatchIndex >= 0) score += (i - lastMatchIndex - 1) * 2;
-			}
-			if (isWordBoundary) score -= 10;
-			score += i * 0.1;
-			lastMatchIndex = i;
-			queryIndex++;
-		}
-	}
-	return queryIndex < queryLower.length ? null : score;
-}
-
-/**
- * Filter items using pre-lowercased searchTextLower field.
- * Supports space-separated tokens (all must match).
- */
-function fuzzyFilterLower<T extends { searchTextLower?: string }>(
-	items: T[],
-	queryLower: string,
-): T[] {
-	const trimmed = queryLower.trim();
-	if (!trimmed) return items;
-
-	const tokens = trimmed.split(/\s+/).filter((t) => t.length > 0);
-	if (tokens.length === 0) return items;
-
-	const results: { item: T; score: number }[] = [];
-	for (const item of items) {
-		const text = item.searchTextLower ?? "";
-		let totalScore = 0;
-		let allMatch = true;
-		for (const token of tokens) {
-			const score = fuzzyMatchLower(token, text);
-			if (score !== null) {
-				totalScore += score;
-			} else {
-				allMatch = false;
-				break;
-			}
-		}
-		if (allMatch) results.push({ item, score: totalScore });
-	}
-	results.sort((a, b) => a.score - b.score);
-	return results.map((r) => r.item);
-}
+import { fuzzyFilterLower, prepareSearchItems } from "./fuzzy-filter.js";
 
 export interface FilterableSelectItem extends SelectItem {
 	/** Additional searchable fields beyond label */
@@ -102,17 +37,9 @@ export class FilterableSelectList implements Component {
 	onCancel?: () => void;
 
 	constructor(items: FilterableSelectItem[], maxVisible: number, theme: FilterableSelectListTheme) {
-		// Pre-compute searchTextLower for each item once
-		this.allItems = items.map((item) => {
-			if (item.searchTextLower) return item;
-			const parts = [item.label];
-			if (item.description) parts.push(item.description);
-			if (item.searchText) parts.push(item.searchText);
-			return { ...item, searchTextLower: parts.join(" ").toLowerCase() };
-		});
+		this.allItems = prepareSearchItems(items);
 		this.maxVisible = maxVisible;
 		this.theme = theme;
-
 		this.input = new Input();
 		this.selectList = new SelectList(this.allItems, maxVisible, theme);
 	}
